Refactor: detach `Illusts` in profile landing page
